+++ app/renderer/pages/devPage/dev.container.js
@@ -4,9 +4,9 @@ import types from 'renderer/types';
 import {
   restartCennzNetNodeChannel,
 } from 'renderer/ipc/cennznet.ipc';
-import type { CennzNetRestartOptions } from '../../../common/types/cennznet-node.types';
+import type { CennzNetRestartOptions } from 'common/types/cennznet-node.types';
 
-const mapStateToProps = ({ dev }) => ({ dev });
+const mapStateToProps = ({ nodeSystem, stream, remoteStream, syncStream, syncRemoteStream }) => ({ nodeSystem, stream, remoteStream, syncStream, syncRemoteStream });
 
 const mapDispatchToProps = dispatch => ({
   onNetworkStatusClick: () => {
@@ -22,7 +22,30 @@ const mapDispatchToProps = dispatch => ({
     };
     restartCennzNetNodeChannel.send(options);
   },
-
+  onStream: () => {
+    dispatch({
+      type: types.stream.requested,
+      payload: {},
+    });
+    dispatch({
+      type: types.remoteStream.requested,
+      payload: {},
+    });
+    dispatch({
+      type: types.syncStream.requested,
+      payload: {},
+    });
+    dispatch({
+      type: types.syncRemoteStream.requested,
+      payload: {},
+    });
+  },
+  onChainSubscribeNewHead: () => {
+    dispatch({
+      type: types.nodeWsChainSubscribeNewHead.requested,
+      payload: {},
+    });
+  },
 
 });