+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Improvements
 
+- Able launch app with runtime parameter
+- Able to launch app with cennznet-node in windows
+- Add ./scripts/init.sh to init dev env
+
 ## v0.2.1 (2019.01.04)
 
 #### Improvements